Add new install instructions
The [latest pull request](https://github.com/lara-press/framework/pull/57) on the framework removes the need for the [larapress](https://github.com/lara-press/larapress) repo. 

These are the instructions for the new [installer](https://github.com/lara-press/installer). This also describes how to manually install with the changes to the framework.


https://github.com/lara-press/framework/tree/alternative-install